fix: wire remaining gaps to real session + APIs
- Login primeiro acesso: 'Salvar e entrar' now calls PUT /api/users/profile
  and PUT /api/users/password in parallel, shows loading/error states
- Vitrine: split into server page (auth + DB fetch) + VitrineClient; Header
  now receives real user name, company, role from DB instead of hardcoded data
- Vitrine detail: split into server page (auth + ebook fetch) + EbookDetailClient;
  Header receives real user data; ebook data is server-rendered (no useEffect fetch)
- auth.ts: add `role` to JWTPayload interface for proper TypeScript typing
